refactor: move restaurant menu fetching into separate function
