+++ pages/index.tsx
@@ -5,15 +5,8 @@ import { utcToZonedTime, format } from 'date-fns-tz';
 
 import { PageTitleWithDate } from '../components/PageTitleWithDate/PageTitleWithDate';
 import { RestaurantOneDayMenu } from '../components/RestaurantOneDayMenu/RestaurantOneDayMenu';
-import { WeekDayEnum } from '../lib/utils/regexUtils';
-import { WeekDay } from '../lib/types/weekdays';
-import { getIsoPajaCurrentDayMenu } from '../lib/utils/getIsoPajaCurrentDayMenu';
-import { getStudio10CurrentDayMenu } from '../lib/utils/getStudio10CurrentDayMenu';
 import { RestaurantMenus } from '../lib/types/restaurantMenus';
 import { WEEKDAYS_ARRAY } from '../lib/constants/weekdaysArray';
-import { scrapeIsoPaja } from '../lib/scrapers/cheerio/scrapeIsoPaja';
-import { scrapeStudio10 } from '../lib/scrapers/cheerio/scrapeStudio10';
-import { fetchHuoltamoCurrentDayMenuFromApi } from '../lib/utils/fetchHuoltamoCurrentDayMenuFromApi';
 import { TIME_ZONE } from '../lib/constants/timeZone';
 import {
   BOX_URL,
@@ -22,9 +15,7 @@ import {
   ISO_PAJA_URL,
   STUDIO_10_URL,
 } from '../lib/constants/restaurantUrls';
-import { scrapeBox } from '../lib/scrapers/cheerio/scrapeBox';
-import { getBoxCurrentDayMenu } from '../lib/utils/getBoxCurrentDayMenu';
-import { fetchDylanCurrentDayMenuFromApi } from '../lib/utils/fetchDylanCurrentDayMenuFromApi';
+import { getAllRestaurantsCurrentDayMenus } from '../lib/utils/getAllRestaurantsCurrentDayMenus';
 
 interface HomeProps {
   restaurant: RestaurantMenus;
@@ -75,20 +66,7 @@ const Home: NextPage<HomeProps> = ({ restaurant, isoDate }) => {
 };
 
 export const getStaticProps: GetStaticProps<HomeProps> = async () => {
-  const currentDay = new Date().toLocaleDateString('default', {
-    weekday: 'long',
-  }) as WeekDay;
-  const isoDate = new Date().toISOString();
-  const weekDayIndex = WeekDayEnum[currentDay];
-  const zonedIsoDate = utcToZonedTime(isoDate, TIME_ZONE);
-
-  const restaurant: RestaurantMenus = {
-    huoltamo: await fetchHuoltamoCurrentDayMenuFromApi(zonedIsoDate),
-    studio10: await getStudio10CurrentDayMenu(weekDayIndex, scrapeStudio10),
-    isoPaja: await getIsoPajaCurrentDayMenu(weekDayIndex, scrapeIsoPaja),
-    box: await getBoxCurrentDayMenu(weekDayIndex, scrapeBox),
-    dylan: await fetchDylanCurrentDayMenuFromApi(weekDayIndex),
-  };
+  const { restaurant, isoDate } = await getAllRestaurantsCurrentDayMenus();
 
   return {
     props: {
